feat: expanding example unit tests
